+++ archon/actor/commands/expose.py
@@ -83,6 +83,14 @@ __all__ = ["expose", "read", "abort"]
     help="Whether to read out the frame.",
 )
 @click.option(
+    "--async-readout",
+    default=False,
+    is_flag=True,
+    help="When set, readout will be initiated but the command returns "
+    "immediately as readout begins. If multiple exposures are commanded only "
+    "the last one will be read out asynchronously.",
+)
+@click.option(
     "--header",
     type=str,
     default="{}",
@@ -132,6 +140,7 @@ async def expose(
     controller: str | None = None,
     flavour: str = "object",
     readout: bool = True,
+    async_readout: bool = False,
     header: str = "{}",
     delay_readout: int = 0,
     count: int = 1,
@@ -170,23 +179,37 @@ async def expose(
 
     for n in range(1, count + 1):
         flavours = [flavour, "dark"] if with_dark else [flavour]
-        for this_flavour in flavours:
+        for n_flavour, this_flavour in enumerate(flavours):
             delegate.use_shutter = not no_shutter
-            result = await delegate.expose(
+            exposure_result = await delegate.expose(
                 command,
                 selected_controllers,
                 flavour=this_flavour,
                 exposure_time=exposure_time,
-                readout=readout,
-                extra_header=extra_header,
-                delay_readout=delay_readout,
+                readout=False,
                 window_mode=window_mode,
                 write=not no_write,
                 seqno=seqno,
             )
 
-            if not result:
+            if not exposure_result:
                 # expose will fail the command.
+                return
+
+            readout_task = asyncio.create_task(
+                delegate.readout(
+                    command,
+                    extra_header=extra_header,
+                    delay_readout=delay_readout,
+                )
+            )
+
+            if async_readout and n == count and n_flavour == len(flavours) - 1:
+                return command.finish("Returning while readout is ongoing.")
+
+            readout_result = await readout_task
+
+            if not readout_result:
                 return
 
     return command.finish()